Openai demo app vertical reorg (#234)
* organize demo ai app vertically

* Update main.wasp.diff

* Update guided-tour.md

* Leftover vertical org changes (#233)

* newsletter dir & leftovers

* update app_diff

* Update guided-tour.md

* Update guided-tour.md
