+++ opensaas-sh/app_diff/src/shared/common.ts.diff
@@ -1,6 +1,7 @@
---- template/app/src/common.ts
-+++ opensaas-sh/app/src/common.ts
+--- template/app/src/shared/common.ts
++++ opensaas-sh/app/src/shared/common.ts
 @@ -1,2 +1,3 @@
  export const DocsUrl = 'https://docs.opensaas.sh';
  export const BlogUrl = 'https://docs.opensaas.sh/blog';
 +export const GithubUrl = 'https://github.com/wasp-lang/open-saas';
+\ No newline at end of file
