+++ template/app/src/shared/common.ts
