+++ template/app/src/shared/utils.ts
